+++ demo/src/main/java/com/orange/ods/demo/ui/components/sliders/ComponentSliders.kt
@@ -19,6 +19,8 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.dimensionResource
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
 import com.orange.ods.compose.component.control.OdsSlider
 import com.orange.ods.compose.component.control.OdsSliderLockups
 import com.orange.ods.demo.R
@@ -46,8 +48,10 @@ fun ComponentSlidersContent() {
             value = discreteWithIconsSliderPosition,
             steps = 10,
             onValueChange = { discreteWithIconsSliderPosition = it },
-            leftIconRes = R.drawable.ic_heart,
-            rightIconRes = R.drawable.ic_heart,
+            leftIcon = painterResource(id = R.drawable.ic_volume_status_1),
+            leftIconContentDescription = stringResource(id = R.string.component_slider_low_volume),
+            rightIcon = painterResource(id = R.drawable.ic_volume_status_4),
+            rightIconContentDescription = stringResource(id = R.string.component_slider_high_volume)
         )
 
         var continuousSliderPosition by remember { mutableStateOf(0f) }
@@ -64,8 +68,10 @@ fun ComponentSlidersContent() {
         OdsSlider(
             value = continuousSliderWithIconsPosition,
             onValueChange = { continuousSliderWithIconsPosition = it },
-            leftIconRes = R.drawable.ic_heart,
-            rightIconRes = R.drawable.ic_heart,
+            leftIcon = painterResource(id = R.drawable.ic_volume_status_1),
+            leftIconContentDescription = stringResource(id = R.string.component_slider_low_volume),
+            rightIcon = painterResource(id = R.drawable.ic_volume_status_4),
+            rightIconContentDescription = stringResource(id = R.string.component_slider_high_volume)
         )
 
         var discreteLockupsSliderPosition by remember { mutableStateOf(0f) }